// src/app/modules/academicSemester/academicSemester.interface.ts
export type TMonth =
  | 'January'
  | 'February'
  | 'March'
  | 'April'
  | 'May'
  | 'June'
  | 'July'
  | 'August'
  | 'September'
  | 'October'
  | 'November'
  | 'December';

export type TAcademicSemesterName = 'Autumn' | 'Summer' | 'Fall';
export type TAcademicSemesterCode = '01' | '02' | '03';

export type TAcademicSemester = {
  name: TAcademicSemesterName;
  code: TAcademicSemesterCode;
  year: string;
  startMonth: TMonth;
  endMonth: TMonth;
};
//semester name ---> semester code
export type TAcademicSemesterNameCodeMapper = {
  // Autumn: '01';
  // Summer: '02';
  // Fall: '03';

  //map type
  [key: string]: string;
};